requirements.txt
Django>=4.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
